requirements.txt
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
Pillow==10.3.0
py3rosmsgs==1.18.2
pycuda==2024.1
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.16.0
sensor-msgs==1.13.1
tensorrt==8.5.2.2
torchvision==0.17.2
